Use common EVT_C_STORE handler
